FISH-492 Refactor imports
