+++ plugins/org.eclipse.payara.tools.ui/src/org/eclipse/payara/tools/ui/micro/wizards/MicroProjectWizardPage.java
@@ -13,7 +13,7 @@ import java.util.regex.Pattern;
 import org.eclipse.m2e.core.project.ProjectImportConfiguration;
 import org.eclipse.m2e.core.ui.internal.wizards.AbstractMavenWizardPage;
 import org.eclipse.m2e.core.ui.internal.wizards.MavenProjectWizardArchetypeParametersPage;
-import org.eclipse.payara.tools.ui.wizards.Messages;
+import org.eclipse.payara.tools.ui.micro.wizards.Messages;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.layout.GridLayout;
