+++ plugins/org.eclipse.payara.tools.ui/src/org/eclipse/payara/tools/ui/micro/wizards/MicroSettingsWizardPage.java
@@ -21,7 +21,7 @@ import java.util.Properties;
 import org.apache.maven.archetype.catalog.Archetype;
 import org.eclipse.m2e.core.project.ProjectImportConfiguration;
 import org.eclipse.m2e.core.ui.internal.wizards.AbstractMavenWizardPage;
-import org.eclipse.payara.tools.ui.wizards.Messages;
+import org.eclipse.payara.tools.ui.micro.wizards.Messages;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.layout.GridLayout;
